fp32 cpu

+++ notebooks/tyler/2023-06-22_librispeech_contrastive.py
@@ -545,9 +545,9 @@ class SpeechOrEMGToText(Model):
 
             # only use vocalized emg for supervised contrastive loss as we have
             # frame-aligned phoneme labels for those
-            z = torch.concatenate([paired_e_z, *audio_z]).to('cpu')
-            z_class = torch.concatenate([*paired_e_phonemes, *audio_phonemes]).to('cpu').float()
-            sup_nce_loss = supervised_contrastive_loss(z, z_class, device='cpu').float()
+            z = torch.concatenate([paired_e_z, *audio_z]).to('cpu').float()
+            z_class = torch.concatenate([*paired_e_phonemes, *audio_phonemes]).to('cpu')
+            sup_nce_loss = supervised_contrastive_loss(z, z_class, device='cpu')
                 # device=self.device)
         elif emg_z is not None:
             # INFO: phoneme labels aren't frame-aligned with emg, so we can't use them
